+++ lib/src/models/common_icon.dart
@@ -4,7 +4,7 @@ class CommonIcon {
   const CommonIcon({
     /// ex: (prefixIcon || suffixIcon) = Icons.done ( runtimeType: TypeIconData)
     ///      (prefixIcon || suffixIcon) = assets/icons/done.svg (runtimeType: String)
-    @required this.path,
+    required this.path,
     this.containerStyle,
     this.color = COMMON_BLACK_COLOR,
     this.iconDataSize = 24.0,